created common library with single h-file called types, which includes uint64 and indexValue

+++ libraries/dataset/src/CompressedIntegerList.cpp
@@ -17,7 +17,7 @@ namespace dataset
 
     void CompressedIntegerList::Iterator::Next()
     {
-        uint delta;
+        uint64 delta;
         _iter += _iter_increment;
         uint8_t first_val = *_iter;
 
@@ -42,7 +42,7 @@ namespace dataset
         _value += delta;
     }
 
-    uint CompressedIntegerList::Iterator::GetValue() const
+    uint64 CompressedIntegerList::Iterator::GetValue() const
     {
         return _value;
     }
@@ -55,17 +55,17 @@ namespace dataset
     CompressedIntegerList::CompressedIntegerList() : _last(UINT64_MAX), _size(0)
     {}
 
-    uint CompressedIntegerList::Size() const
+    uint64 CompressedIntegerList::Size() const
     {
         return _size;
     }
 
-    void CompressedIntegerList::Reserve(uint size)
+    void CompressedIntegerList::Reserve(uint64 size)
     {
         _mem.reserve(size*2); // guess that, on average, every entry will occupy 2 bytes
     }
 
-    uint CompressedIntegerList::Max() const
+    uint64 CompressedIntegerList::Max() const
     {
         if(_size == 0)
         {
@@ -76,12 +76,12 @@ namespace dataset
     }
 
     /// adds an integer at the end of the list
-    void CompressedIntegerList::PushBack(uint value)
+    void CompressedIntegerList::PushBack(uint64 value)
     {
         assert(value != UINT64_MAX);    // special value reserved for initialization
 
         // calculate the delta from the previous number pushed
-        uint delta = 0;
+        uint64 delta = 0;
 
         // allow the first PushBack to have a value of zero, but subsequently require an increasing value
         if (_last < UINT64_MAX)
@@ -122,7 +122,7 @@ namespace dataset
         unsigned int mask = log2bytes << 6; // mask == log2(# bytes) shifted to be high-order 2 bits of a byte
         // splice the data length encoding in as the top 2 bits of the first byte
         // So, move all high-order bits of the delta over by 2 to make room, Add the mask, and Add the residual low-order bits of the delta
-        uint write_val = ((delta << 2) & 0xffffffffffffff00) | mask | (delta & 0x3f);
+        uint64 write_val = ((delta << 2) & 0xffffffffffffff00) | mask | (delta & 0x3f);
         memcpy(buf, &write_val, total_bytes);
 
         ++_size;
