+++ libraries/dataset/src/SparseBinaryDatavector.cpp
@@ -19,7 +19,7 @@ namespace dataset
     }
 
     template<typename IntegerListType>
-    uint SparseBinaryDataVectorBase<IntegerListType>::Iterator::GetIndex() const
+    uint64 SparseBinaryDataVectorBase<IntegerListType>::Iterator::GetIndex() const
     {
         return _list_iterator.GetValue();
     }
@@ -40,7 +40,7 @@ namespace dataset
     {}
 
     template<typename IntegerListType>
-    void SparseBinaryDataVectorBase<IntegerListType>::PushBack(uint index, double value)
+    void SparseBinaryDataVectorBase<IntegerListType>::PushBack(uint64 index, double value)
     {
         if (value == 0)
         {
@@ -59,18 +59,18 @@ namespace dataset
     }
 
     //template<typename IntegerListType>
-    //void SparseBinaryDataVectorBase<IntegerListType>::foreach_nonzero(function<void(uint, double)> func, uint index_offset) const // TODO
+    //void SparseBinaryDataVectorBase<IntegerListType>::foreach_nonzero(function<void(uint64, double)> func, uint64 index_offset) const // TODO
     //{
     //    auto iter = _indices.GetIterator();
     //    while(iter.IsValid())
     //    {
-    //        func(index_offset + iter.GetValue(), (double)1.0);
+    //        func(index_offset + iter.value, (double)1.0);
     //        iter.Next();
     //    }
     //}
 
     template<typename IntegerListType>
-    uint SparseBinaryDataVectorBase<IntegerListType>::Size() const
+    uint64 SparseBinaryDataVectorBase<IntegerListType>::Size() const
     {
         if(_indices.Size() == 0)
         {
@@ -83,7 +83,7 @@ namespace dataset
     }
 
     template<typename IntegerListType>
-    uint SparseBinaryDataVectorBase<IntegerListType>::NumNonzeros() const
+    uint64 SparseBinaryDataVectorBase<IntegerListType>::NumNonzeros() const
     {
         return _indices.Size();
     }
