+++ libraries/linear/src/DoubleMatrix.cpp
@@ -8,17 +8,17 @@ using std::fill;
 
 namespace linear
 {
-    uint DoubleMatrixBase::NumRows() const
+    uint64 DoubleMatrixBase::NumRows() const
     {
         return _num_rows;
     }
 
-    uint DoubleMatrixBase::NumColumns() const
+    uint64 DoubleMatrixBase::NumColumns() const
     {
         return _num_columns;
     }
 
-    void DoubleMatrixBase::Set(uint i, uint j, double value)
+    void DoubleMatrixBase::Set(uint64 i, uint64 j, double value)
     {
         operator()(i,j) = value;
     }
@@ -30,7 +30,7 @@ namespace linear
 
     void DoubleMatrixBase::Gemv(const double* p_x, double* p_y, double alpha, double beta) const
     {
-        int size = (int)NumRows(); // openmp doesn't like uint
+        int size = (int)NumRows(); // openmp doesn't like uint64
 
         if(alpha == 1.0)
         {
@@ -90,7 +90,7 @@ namespace linear
 
     void DoubleMatrixBase::Gevm(const double* p_x, double* p_y, double alpha, double beta) const
     {
-        int size = (int)NumColumns(); // openmp doesn't like uint
+        int size = (int)NumColumns(); // openmp doesn't like uint64
 
         if (alpha == 1.0)
         {
@@ -150,9 +150,9 @@ namespace linear
 
     void DoubleMatrixBase::Print(ostream& os) const
     {
-        for(uint i = 0; i < _num_rows; ++i)
+        for(uint64 i = 0; i < _num_rows; ++i)
         {
-            for(uint j = 0; j < _num_columns; ++j)
+            for(uint64 j = 0; j < _num_columns; ++j)
             {
                 os << this->operator()(i, j) << "\t";
             }
@@ -160,24 +160,24 @@ namespace linear
         }
     }
 
-    DoubleMatrixBase::DoubleMatrixBase(uint size, uint numRows, uint numColumns)
+    DoubleMatrixBase::DoubleMatrixBase(uint64 size, uint64 numRows, uint64 numColumns)
         : vector<double>(size), _num_rows(numRows), _num_columns(numColumns)
     {}
 
-    double DoubleMatrixBase::RowDot(uint i, const double* p_x) const
+    double DoubleMatrixBase::RowDot(uint64 i, const double* p_x) const
     {
         double value = 0;
-        for (uint j = 0; j < NumColumns(); ++j)
+        for (uint64 j = 0; j < NumColumns(); ++j)
         {
             value += (double)(this->operator()(i, j) * p_x[j]);
         }
         return value;
     }
 
-    double DoubleMatrixBase::ColumnDot(uint j, const double* p_x) const
+    double DoubleMatrixBase::ColumnDot(uint64 j, const double* p_x) const
     {
         double value = 0;
-        for (uint i = 0; i < NumRows(); ++i)
+        for (uint64 i = 0; i < NumRows(); ++i)
         {
             value += (double)(this->operator()(i, j) * p_x[i]);
         }
@@ -186,17 +186,17 @@ namespace linear
 
     // MatrixStructure::column
 
-    DoubleMatrix<MatrixStructure::column>::DoubleMatrix(uint numRows, uint numColumns) :
+    DoubleMatrix<MatrixStructure::column>::DoubleMatrix(uint64 numRows, uint64 numColumns) :
         DoubleMatrixBase(numRows*numColumns, numRows, numColumns)
     {}
 
-    double& DoubleMatrix<MatrixStructure::column>::operator()(uint i, uint j)
+    double& DoubleMatrix<MatrixStructure::column>::operator()(uint64 i, uint64 j)
     {
         assert(i < this->_num_rows && j < this->_num_columns);
         return this->operator[](j * this->_num_rows + i);
     }
 
-    double DoubleMatrix<MatrixStructure::column>::operator()(uint i, uint j) const
+    double DoubleMatrix<MatrixStructure::column>::operator()(uint64 i, uint64 j) const
     {
         assert(i < this->_num_rows && j < this->_num_columns);
         return this->operator[](j * this->_num_rows + i);
@@ -204,17 +204,17 @@ namespace linear
 
     // MatrixStructure::row
 
-    DoubleMatrix<MatrixStructure::row>::DoubleMatrix(uint numRows, uint numColumns) :
+    DoubleMatrix<MatrixStructure::row>::DoubleMatrix(uint64 numRows, uint64 numColumns) :
         DoubleMatrixBase(numRows*numColumns, numRows, numColumns)
     {}
 
-    double& DoubleMatrix<MatrixStructure::row>::operator()(uint i, uint j)
+    double& DoubleMatrix<MatrixStructure::row>::operator()(uint64 i, uint64 j)
     {
         assert(i < this->_num_rows && j < this->_num_columns);
         return this->operator[](i * this->_num_columns + j);
     }
 
-    double DoubleMatrix<MatrixStructure::row>::operator()(uint i, uint j) const
+    double DoubleMatrix<MatrixStructure::row>::operator()(uint64 i, uint64 j) const
     {
         assert(i < this->_num_rows && j < this->_num_columns);
         return this->operator[](i * this->_num_columns + j);
@@ -222,23 +222,23 @@ namespace linear
 
     // MatrixStructure::column_square
 
-    DoubleMatrix<MatrixStructure::column_square>::DoubleMatrix(uint dimension) :
+    DoubleMatrix<MatrixStructure::column_square>::DoubleMatrix(uint64 dimension) :
         DoubleMatrix<MatrixStructure::column>(dimension, dimension)
     {}
 
     // MatrixStructure::row_square
 
-    DoubleMatrix<MatrixStructure::row_square>::DoubleMatrix(uint dimension) :
+    DoubleMatrix<MatrixStructure::row_square>::DoubleMatrix(uint64 dimension) :
         DoubleMatrix<MatrixStructure::row>(dimension, dimension)
     {}
 
     // MatrixStructure::row_square_uptriangular
 
-    DoubleMatrix<MatrixStructure::row_square_uptriangular>::DoubleMatrix(uint dimension) :
+    DoubleMatrix<MatrixStructure::row_square_uptriangular>::DoubleMatrix(uint64 dimension) :
         DoubleMatrixBase(dimension*(dimension + 1) / 2, dimension, dimension)
     {}
 
-    void DoubleMatrix<MatrixStructure::row_square_uptriangular>::Set(uint i, uint j, double value)
+    void DoubleMatrix<MatrixStructure::row_square_uptriangular>::Set(uint64 i, uint64 j, double value)
     {
         if(i > j)
         {
@@ -248,7 +248,7 @@ namespace linear
         DoubleMatrixBase::Set(i, j, value);
     }
 
-    double& DoubleMatrix<MatrixStructure::row_square_uptriangular>::operator()(uint i, uint j)
+    double& DoubleMatrix<MatrixStructure::row_square_uptriangular>::operator()(uint64 i, uint64 j)
     {
         assert(i < this->_num_rows && j < this->_num_columns);
         _dummy = 0;
@@ -256,7 +256,7 @@ namespace linear
         return this->operator[](i * this->_num_columns - (i + 1) * i / 2 + j);
     }
 
-    double DoubleMatrix<MatrixStructure::row_square_uptriangular>::operator()(uint i, uint j) const
+    double DoubleMatrix<MatrixStructure::row_square_uptriangular>::operator()(uint64 i, uint64 j) const
     {
         assert(i < this->_num_rows && j < this->_num_columns);
         if(i > j) return 0;
@@ -277,11 +277,11 @@ namespace linear
 
     // MatrixStructure::diagonal
 
-    DoubleMatrix<MatrixStructure::diagonal>::DoubleMatrix(uint dimension) :
+    DoubleMatrix<MatrixStructure::diagonal>::DoubleMatrix(uint64 dimension) :
         DoubleMatrixBase(dimension, dimension, dimension)
     {}
 
-    void DoubleMatrix<MatrixStructure::diagonal>::Set(uint i, uint j, double value)
+    void DoubleMatrix<MatrixStructure::diagonal>::Set(uint64 i, uint64 j, double value)
     {
         if(i != j)
         {
@@ -291,7 +291,7 @@ namespace linear
         DoubleMatrixBase::Set(i, j, value);
     }
 
-    double& DoubleMatrix<MatrixStructure::diagonal>::operator()(uint i, uint j)
+    double& DoubleMatrix<MatrixStructure::diagonal>::operator()(uint64 i, uint64 j)
     {
         assert(i < this->_num_rows && j < this->_num_columns);
         _dummy = 0;
@@ -299,7 +299,7 @@ namespace linear
         return this->operator[](i);
     }
 
-    double DoubleMatrix<MatrixStructure::diagonal>::operator()(uint i, uint j) const
+    double DoubleMatrix<MatrixStructure::diagonal>::operator()(uint64 i, uint64 j) const
     {
         assert(i < this->_num_rows && j < this->_num_columns);
         if(i != j) return 0;
